refactor(console): move test files to co-located style (drop __tests__ dirs)
Place *.test.{ts,tsx} files directly alongside their source files,
following the openclaw convention and standard Vitest/React practice.
Removes 18 __tests__/ subdirectories.
