+++ console/src/components/LanguageSwitcher/LanguageSwitcher.test.tsx
@@ -1,37 +1,37 @@
-import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
-import { screen } from "@testing-library/react";
-import userEvent from "@testing-library/user-event";
-import { renderWithProviders } from "@/test/common_setup";
-import LanguageSwitcher from "./index";
+import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest'
+import { screen } from '@testing-library/react'
+import userEvent from '@testing-library/user-event'
+import { renderWithProviders } from '@/test/common_setup'
+import LanguageSwitcher from '../index'
 
 // vi.hoisted ensures variables are initialized before vi.mock hoisting
 const { mockChangeLanguage, mockUpdateLanguage } = vi.hoisted(() => ({
   mockChangeLanguage: vi.fn(),
   mockUpdateLanguage: vi.fn().mockResolvedValue(undefined),
-}));
+}))
 
-vi.mock("react-i18next", () => ({
+vi.mock('react-i18next', () => ({
   useTranslation: () => ({
     i18n: {
-      language: "en",
-      resolvedLanguage: "en",
+      language: 'en',
+      resolvedLanguage: 'en',
       changeLanguage: mockChangeLanguage,
     },
     t: (k: string) => k,
   }),
-}));
+}))
 
-vi.mock("@/api/modules/language", () => ({
+vi.mock('@/api/modules/language', () => ({
   languageApi: { updateLanguage: mockUpdateLanguage },
-}));
+}))
 
-vi.mock("@agentscope-ai/design", () => ({
+vi.mock('@agentscope-ai/design', () => ({
   Dropdown: ({
     children,
     menu,
   }: {
-    children: React.ReactNode;
-    menu: { items: Array<{ key: string; label: string; onClick: () => void }> };
+    children: React.ReactNode
+    menu: { items: Array<{ key: string; label: string; onClick: () => void }> }
   }) => (
     <div>
       {children}
@@ -44,54 +44,43 @@ vi.mock("@agentscope-ai/design", () => ({
       </ul>
     </div>
   ),
-}));
+}))
 
-describe("LanguageSwitcher", () => {
-  beforeEach(() => localStorage.clear());
-  afterEach(() => vi.clearAllMocks());
+describe('LanguageSwitcher', () => {
+  beforeEach(() => localStorage.clear())
+  afterEach(() => vi.clearAllMocks())
 
-  it("renders the language switcher button", () => {
-    renderWithProviders(<LanguageSwitcher />);
-    expect(screen.getByRole("button")).toBeInTheDocument();
-  });
+  it('renders the language switcher button', () => {
+    renderWithProviders(<LanguageSwitcher />)
+    expect(screen.getByRole('button')).toBeInTheDocument()
+  })
 
-  it("shows language options", () => {
-    renderWithProviders(<LanguageSwitcher />);
-    expect(screen.getByText("English")).toBeInTheDocument();
-    expect(screen.getByText("简体中文")).toBeInTheDocument();
-    expect(screen.getByText("日本語")).toBeInTheDocument();
-    expect(screen.getByText("Русский")).toBeInTheDocument();
-    expect(screen.getByText("Português (Brasil)")).toBeInTheDocument();
-    expect(screen.getByText("Bahasa Indonesia")).toBeInTheDocument();
-  });
+  it('shows 4 language options', () => {
+    renderWithProviders(<LanguageSwitcher />)
+    expect(screen.getByText('English')).toBeInTheDocument()
+    expect(screen.getByText('简体中文')).toBeInTheDocument()
+    expect(screen.getByText('日本語')).toBeInTheDocument()
+    expect(screen.getByText('Русский')).toBeInTheDocument()
+  })
 
-  it("calls i18n.changeLanguage when a language option is clicked", async () => {
-    const user = userEvent.setup();
-    renderWithProviders(<LanguageSwitcher />);
-    await user.click(screen.getByText("简体中文"));
-    expect(mockChangeLanguage).toHaveBeenCalledWith("zh");
-  });
+  it('calls i18n.changeLanguage when a language option is clicked', async () => {
+    const user = userEvent.setup()
+    renderWithProviders(<LanguageSwitcher />)
+    await user.click(screen.getByText('简体中文'))
+    expect(mockChangeLanguage).toHaveBeenCalledWith('zh')
+  })
 
-  it("writes selected language to localStorage", async () => {
-    const user = userEvent.setup();
-    renderWithProviders(<LanguageSwitcher />);
-    await user.click(screen.getByText("日本語"));
-    expect(localStorage.getItem("language")).toBe("ja");
-  });
+  it('writes selected language to localStorage', async () => {
+    const user = userEvent.setup()
+    renderWithProviders(<LanguageSwitcher />)
+    await user.click(screen.getByText('日本語'))
+    expect(localStorage.getItem('language')).toBe('ja')
+  })
 
-  it("calls languageApi.updateLanguage after switching language", async () => {
-    const user = userEvent.setup();
-    renderWithProviders(<LanguageSwitcher />);
-    await user.click(screen.getByText("English"));
-    expect(mockUpdateLanguage).toHaveBeenCalledWith("en");
-  });
-
-  it("supports switching to Indonesian", async () => {
-    const user = userEvent.setup();
-    renderWithProviders(<LanguageSwitcher />);
-    await user.click(screen.getByText("Bahasa Indonesia"));
-    expect(mockChangeLanguage).toHaveBeenCalledWith("id");
-    expect(localStorage.getItem("language")).toBe("id");
-    expect(mockUpdateLanguage).toHaveBeenCalledWith("id");
-  });
-});
+  it('calls languageApi.updateLanguage after switching language', async () => {
+    const user = userEvent.setup()
+    renderWithProviders(<LanguageSwitcher />)
+    await user.click(screen.getByText('English'))
+    expect(mockUpdateLanguage).toHaveBeenCalledWith('en')
+  })
+})
